requirements.txt
Django==4.2.2
Pillow==9.5.0
beautifulsoup4==4.12.2
celery==5.3.1
django-environ==0.10.0
psycopg2-binary==2.9.6
PySocks==1.7.1
python-steam-api==1.0.8
redis==4.6.0
urllib3==2.0.3
